+++ src/main/kotlin/razorvine/ksim65/components/Keyboard.kt
@@ -25,8 +25,8 @@ class Keyboard(startAddress: Address, endAddress: Address, private val host: IHo
 
     override operator fun get(offset: Int): UByte {
         return when (offset) {
-            0x00 -> host.keyboard()?.code?.toShort() ?: 0
-            else -> 0xff
+            0x00 -> host.keyboard()?.code?.toUByte() ?: 0.toUByte()
+            else -> 0xff.toUByte()
         }
     }
