+++ src/test/kotlin/TestC64Components.kt
@@ -9,44 +9,44 @@ class TestC64Components {
         val io = CpuIoPort()
         io.reset()
 
-        assertEquals(0xef, io[0])
-        assertEquals(0x37, io[1])
+        assertEquals(0xef.toUByte(),  io[0])
+        assertEquals(0x37.toUByte(),  io[1])
 
         io[0] = 0b11111111
         io[1] = 0b11111111
-        assertEquals(0b11111111,io[0])
-        assertEquals(0b11111111,io[1])
+        assertEquals(0b11111111.toUByte(), io[0])
+        assertEquals(0b11111111.toUByte(), io[1])
         io[0] = 0
-        assertEquals(0b11011111,io[1], "bit 5 (cassette motor) is drawn low if input")
+        assertEquals(0b11011111.toUByte(), io[1], "bit 5 (cassette motor) is drawn low if input")
         io[0] = 0b11111111
-        assertEquals(0b11111111,io[1], "bit 5 (cassette motor) is high if output")
+        assertEquals(0b11111111.toUByte(), io[1], "bit 5 (cassette motor) is high if output")
 
         io[1] = 0x37
         io[0] = 0b00100000
-        assertEquals(0b00100000, io[0])
-        assertEquals(0x37, io[1])
+        assertEquals(0b00100000.toUByte(),  io[0])
+        assertEquals(0x37.toUByte(),  io[1])
 
         io[0] = 0xff
-        assertEquals(0xff, io[0])
-        assertEquals(0x37, io[1])
+        assertEquals(0xff.toUByte(),  io[0])
+        assertEquals(0x37.toUByte(),  io[1])
 
         io[1] = 0b11110000
-        assertEquals(0b11110000, io[1])
+        assertEquals(0b11110000.toUByte(),  io[1])
 
         io[0] = 0b00100000
         io[1] = 0b01010101
-        assertEquals(0b11010000, io[1])
+        assertEquals(0b11010000.toUByte(),  io[1])
         io[0] = 0b00100001
-        assertEquals(0b11010000, io[1])
+        assertEquals(0b11010000.toUByte(),  io[1])
         io[1] = 0b01010101
-        assertEquals(0b11010001, io[1])
+        assertEquals(0b11010001.toUByte(),  io[1])
 
         io[0] = 0b11000011
-        assertEquals(0b11010001, io[1])
+        assertEquals(0b11010001.toUByte(),  io[1])
         io[1] = 0b00100000
-        assertEquals(0b00010000, io[1])
+        assertEquals(0b00010000.toUByte(),  io[1])
         io[1] = 0b11111111
-        assertEquals(0b11010011, io[1])
+        assertEquals(0b11010011.toUByte(),  io[1])
     }
 
     @Test
